+++ packages/@react-spectrum/tabs/stories/Tabs.stories.js
@@ -18,33 +18,51 @@ import {storiesOf} from '@storybook/react';
 import {Tabs} from '..';
 
 storiesOf('Tabs', module)
-  .add(
-    'Default',
-    () => render()
+  .add('Default', () => render(), {
+    info:
+      'The page, anchored, and panel variants have been deprecated. Page is now compact, and panel/anchored tabs are just default. These variants will map properly to the new styles, but please do not specify these variants going forward.'
+  })
+  .add('defaultSelectedItem: val2', () =>
+    render({defaultSelectedItem: 'val2'})
   )
-  .add(
-    'orientation = vertical',
-    () => render({orientation: 'vertical'})
+  .add('orientation: vertical', () => render({orientation: 'vertical'}))
+  .add('density: compact', () => render({density: 'compact'}))
+  .add('isQuiet', () => render({isQuiet: true}))
+  .add('isQuiet, density: compact', () =>
+    render({isQuiet: true, density: 'compact'})
   )
-  .add(
-    'orientation = vertical, icons',
-    () => (
-      <Tabs orientation="vertical" onSelect={action('onSelect')}>
-        <Item icon={<Bell />} key="Bell1" title="Bell">Tab Body 1</Item>
-        <Item icon={<Bell />} key="Bell2" title="Bell">Tab Body 2</Item>
-      </Tabs>
-    )
+  .add('icons', () => render({icons: true}))
+  .add('icons, orientation: vertical', () =>
+    render({icons: true, orientation: 'vertical'})
   )
-  .add(
-    'selectedItem = val2 (controlled)',
-    () => render({selectedItem: 'val2'})
+  .add('density: compact, orientation: vertical', () =>
+    render({density: 'compact', orientation: 'vertical'})
   )
-  .add(
-    'defaultSelectedItem = val2 (uncontrolled)',
-    () => render({defaultSelectedItem: 'val2'})
-  )
-  .add('keyboardActivation = manual',
-    () => render({keyboardActivation: 'manual'})
+  // .add(
+  //   'icons only',
+  //   () => (
+  //     <TabList onSelectionChange={action('onSelectionChange')}>
+  //       <Tab icon={<Twitter />} title="Twitter" aria-title="Tab 1" />
+  //       <Tab icon={<Instagram />} title="Instagram" aria-title="Tab 2" />
+  //       <Tab icon={<Facebook />} title="Facebook" aria-title="Tab 3" />
+  //     </TabList>
+  //   ),
+  //   {inline: true}
+  // )
+  // .add(
+  //   'disabled tabs',
+  //   () => (
+  //     <TabList onSelectionChange={action('onSelectionChange')}>
+  //       <Tab icon={<Twitter />} title="Twitter">Tab 1</Tab>
+  //       <Tab icon={<Instagram />} title="Instagram" disabled>Tab 2</Tab>
+  //       <Tab icon={<Facebook />} title="Facebook">Tab 3</Tab>
+  //     </TabList>
+  //   ),
+  //   {inline: true}
+  // )
+  .add('disable all tabs', () => render({isDisabled: true}))
+  .add('keyboardActivation: manual', () =>
+    render({icons: true, keyboardActivation: 'manual'})
   );
 
 function render(props = {}) {
@@ -52,6 +70,7 @@ function render(props = {}) {
     <Tabs {...props} onSelectionChange={action('onSelectionChange')}>
       <Item title="Tab 1" key="val1">Tab Body 1</Item>
       <Item title="Tab 2" key="val2">Tab Body 2</Item>
+      <Item title="Tab 3" key="val3">Tab Body 3</Item>
     </Tabs>
   );
 }
