+++ packages/@react-types/tabs/README.md
@@ -1,0 +1,3 @@
+# @react-types/tabs
+
+This package is part of [react-spectrum](https://github.com/adobe/react-spectrum). See the repo for more details.
